Move BufVec into musli crate
